# src/main/kotlin/com/marcelomachado/urlshortener/controller/openapi.yaml
openapi: 3.0.0
info:
  title: URL Shortener API
  description: API para encurtador de URLs
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /url-shortener:
    post:
      tags:
        - Sync
      summary: Cria uma nova URL e retorna de forma síncrona
      description: Cria uma nova URL e retorna de forma síncrona
      parameters:
        - in: query
          name: url
          schema:
            type: string
          required: true
          description: URL a ser encurtada
      responses:
        201:
          description: URL encurtada criada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: URL encurtada
              examples:
                example1:
                  summary: URL encurtada
                  value:
                    url: "http://short.url/abc123"
        400:
          description: URL inválida, por favor, verifique o formato e tente novamente.

    get:
      tags:
        - Sync
      summary: Retorna a URL salva
      description: Retorna a URL salva
      parameters:
        - in: query
          name: url
          schema:
            type: string
          required: true
          description: URL salva
      responses:
        200:
          description: URL salva retornada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: URL salva

  /url-shortener/kafka:
    post:
      tags:
        - Async
      summary: Processa a URL de forma assíncrona
      description: Encaminha a URL para Kafka
      parameters:
        - in: query
          name: url
          schema:
            type: string
          required: true
          description: URL a ser encurtada
      responses:
        202:
          description: URL encaminhada para Kafka com sucesso
        400:
          description: URL inválida, por favor, verifique o formato e tente novamente.